pyproject.toml
[project]
name = "x-cli"
version = "0.1.0"
description = "CLI for X/Twitter API v2"
requires-python = ">=3.12"
dependencies = [
    "click>=8.3.1",
    "httpx>=0.28.1",
    "rich>=14.3.3",
    "python-dotenv>=1.2.2",
]

[project.scripts]
x-cli = "x_cli.cli:main"

[build-system]
requires = ["hatchling>=1.29.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/x_cli"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = ["pytest>=9.0.2", "ruff>=0.15.5"]

[tool.uv]
package = false
exclude-newer = "7 days"
